Starting off easy
There is most likely going to be an unreal amount of commits, so buckle in.
My idea for the game (currently - might end up getting trashed depending on how it goes) is a Clue inspired murder mystery type game. Very original, right? The player is a detective and the goal is to collect enough evidence for a conviction.
In its current form, the map is basically a direct rip off of the Clue movie set - one house, three levels, with the first floor having the majority of the rooms.
Code wise, I've created all the rooms and added an Item and Person class, because I'm pretty sure I'm gonna need those, eventually.