A more concrete idea
I think I have the idea for my game a bit more hashed out.
I wanted to have some fourth wall breaks and  have some player decisions change the game storyline or gameplay in some way, so I introduced a trapped NPC that plays the role of some player that hasn't figured out how to win the game, and is therefore stuck in the game.
The player can choose to work with this NPC or work alone, and I want that decision to affect things later on.
The new class Game2 is being used to test that idea. The game intro where the player chooses whether or not to help the player is mostly hashed out. That's about it right now, though.