Figuring out items and inventory.
Hashed out the Item Class a bit more.
Removed the previous hashMap and instead added a local Room variable that will hold what room the item is in.
Added methods to: return what room the item is in, change the item status to found, and add item to person inventory.
Started on the Person class. Right now every person, both player and npc can have an inventory. Which currently is just an arrayList of Items.
Added the method dropItem which removes the item from the inventory and adds it to the current room the person is in.