More work on inventory
Added an addItem() method that adds a valid item to a person's inventory if there is room.
Changed the local person variable in the Item class to a string that holds the name of the person who has the item, rather that the person object itself. Might change that later though.