Allowing NPC's to move rooms.
I'm jumping around terribly. I'll try to make these commits as clear as possible so it doesn't read like train of thought, but we'll see how successful I am.
I'm running into a bit of a problem with the person class.
I wanted to move the currentRoom field to be handled by the person class that way any person object has a currentRoom. But I'm running into issues implementing that.
I wanted all characters (player included) to use this class, but I'm in the slight pickle that NPC's and the player don't move the same. NPC's can basically teleport to where I need them to go, but a player can't. Also I want error messages to print if the player tries to do something (add to inventory, drop item, or go to room) that isn't valid, but not if an NPC tries it and it fails for whatever reason.