Modifying the game2 class
Moved the initialization of the rooms, characters, and items to the top of the class, because I'm pretty sure I'm going to need to call them in various methods.
Also added the command GIVE to the CommandWord class that will eventually allow the player to give inventory items to NPCs.