Changing goRoom() responsibility
Added goRoom() to Person class which now handles checking if an exit is null and printing out approriate error messages. Also takes care of moving the player to the new room and printing out room description.
In Game2 class: sketched out how I want help command to work. I'd like the help command to show the objective and print out the command words and what they are used for. So I split the old printHelp() into printObjective() and printCommands().