Moving items around
In the Room class: added two methods. one to add an item to the room's arrayList for items and one to remove it.
In the Item class: modified the addToRoom() and addToInventory() methods to also call to add/ remove the item from the room's arrayList.
In the Person class: modified the addToInventory and addToRoom parameter calls. And I added a print statement that notifies the player that an item has been dropped or that the dropped failed.
In the Game2 class: added some more rooms.