Little fixes to quit command
I was starting to tackle the script/ actual gameplay part of this, but I realized that, currently, the game didn't double check to make sure the player wanted to quit, before quiting so I fixed that.
Other than that, in the Game2 class: I started writing the game intro/ exposition.
In the Room class: I added a method to remove an exit from the room.