Finally playing
So, I think I've laid down as much framework as I'm able to, so now we're tackling the actual gameplay part.
Almost all changes are aesthetic ones: modifying what's printed out to the terminal and tweaking the pacing of the prints.
In the Game2 class: I changed how the quit method so it works with the code I've written so far.
In the Person class: I realized that I never coded an assignment to the name variable, so I fixed that by passing a string parameter to the constructor and added a accessor method for name.
In the Room class: cosmetic changes to how the exitString prints out.