Locked rooms
In Room class: added a locked variable and accessor and mutator methods for it. The variable is assigned with a parameter passed to the constructor.
In the Person class: in the goRoom() method I added a conditional test to check if the room the player is trying to go to is locked. If it is, a message is printed.
In the Game2 class and Game class: modified all room initializations to include locked status.