More jumping around
In Game2 class: Finally gave the rooms detailed descriptions that are printed out when the player first enters the room, which took an ungodly amount of time. Props to writers, we appreciate your sacrifice. I also added a few new rooms. Added all the items into the game.
In the item class: I added another constructor for items inside other items. (e.g. a key inside a box) I have a small issue where item location isn't streamlined, so each item has a room it belongs to (local variable in the item class) and the room has an arrayList of the items in it (in the room class). Right now those needed to be assigned separately. I'll fix that if I have time. Also added an accessor method that returns the room of the item that an item is in (e.g. returns the box's room  for the key)
In the Person class: I sort of have a goBack() method, but I haven't tested it, and I'm not sure I like it. We'll see. I also created a giveItem() method that lets items be exchanged from inventories.
In the Room class: Add the exitString to the end of the detailed room description.