Refactoring? plus more becasue I forgot to commit the last stage.
In Item Class: I realized the item class was doing way too much. I removed the person and room variables, removed the addToInventory() method and addToRoom() method. Added more items  -.-
In the Person Class: Changed the inventory from an ArrayList to a HashMap to make it easier to do other things (like drop item and add items). Changed the inventory system from slot-based to weight based and added the "magic cookie" from textbook in the form of a backpack. Put all responsibility of removing or adding items to rooms or inventories on the Person class methods.
In Room Class: also changed the items arrayList to a HashMap. Added a method to check if an item is in the room (hasItem() ) and one to return an item's object given it's string name.