Refactoring some more
In Item class: Tweaked the constructor, it now assigns found and usable variables. I added an ArrayList of items that holds any items inside the item. (e.g. keys in a box) Also added a usable boolean variable that determines whether or not an item can be use. Added an accessor method for usable status. Added add() method that adds an item (key) to the item's (box) arrayList of items.
In the Person Class: started added the command methods like examineItem(), useItem() that handle checking if an item is in the room/ inventory and print out appropriate messages. Added local variable decked that will serve as a rudimentary health system at some point.
In the Room Class: Changed the ArrayList holding people to a HashMap and corresponding accessor and mutator methods for it.
In the Game2 Class: Fixed all the items that my changes to the item class broke. Started hashing out how commandwords will work.
In the Parser and Command Classes: I added a fouth word option, so the parser can pick up four words on the input line. Fix/ added all the necessary accessors that this change required.