Use command finally done.
In the Game2 class: I tackled the useItem() command and I conquered...I hope. I haven't tested it yet, but I think it looks good and it took 10 years off my life. The useItem() method uses a whole bunch of switch cases to determine what item the player is trying to use and then executes the corresponding code. Finally getting around to NPCs, too.
In the Item class: added a locked variable that allows for container items to be checked if locked before showing what items are inside.
In the Person Class: I had a lot of fun here. Added two boolean variables: dirty and conscious and their accessors and mutators. Concious serves as a death switch. If the player gets decked too many times (accumulates too many injuries) they pass out and it's game over. Dirty is more of a state variable, if true the NPC's won't speak with the player. Added methods to heal up . Added hasItem() which serves to check inventory for an item.
In the Room class: added a boolean variable blocked which is only used to determine wheter player has a light to get through the secret passages. Also added isEmpty() to check if there are any NPCs in the room.